Push the YouTube embed into a Liquid template; add a title
