+++ src/_includes/embeds/youtube.html
@@ -1,0 +1,15 @@
+{% assign video_id = include.url | split: '=' | last %}
+
+<iframe
+  class="youtube"
+  id="youtube_{{ video_id }}"
+  src="https://www.youtube-nocookie.com/embed/{{ video_id }}"
+  title="{{ include.title }}"
+  credentialless
+  sandbox="allow-scripts allow-same-origin"
+  allow="accelerometer 'none'; ambient-light-sensor 'none'; autoplay 'none'; battery 'none'; bluetooth 'none'; browsing-topics 'none'; camera 'none'; ch-ua 'none'; display-capture 'none'; domain-agent 'none'; document-domain 'none'; encrypted-media 'none'; execution-while-not-rendered 'none'; execution-while-out-of-viewport 'none'; gamepad 'none'; geolocation 'none'; gyroscope 'none'; hid 'none'; identity-credentials-get 'none'; idle-detection 'none'; keyboard-map 'none'; local-fonts 'none'; magnetometer 'none'; microphone 'none'; midi 'none'; navigation-override 'none'; otp-credentials 'none'; payment 'none'; picture-in-picture 'none'; publickey-credentials-create 'none'; publickey-credentials-get 'none'; screen-wake-lock 'none'; serial 'none'; speaker-selection 'none'; sync-xhr 'none'; usb 'none'; web-share 'none'; window-management 'none'; xr-spatial-tracking 'none'"
+  csp="sandbox allow-scripts allow-same-origin;"
+  style="aspect-ratio: 560 / 315; width: 100%; max-width: 560px;"
+  frameborder="0"
+  allowfullscreen
+  loading="lazy"></iframe>
