+++ src/_plugins/atom_feed_filters.rb
@@ -81,6 +81,24 @@ module Jekyll
       doc.to_s.gsub('<p></p>', '')
     end
 
+    # According to https://github.com/rubys/feedvalidator, embedding
+    # <iframe> in an RSS feed can be a security risk, so instead we replace
+    # such YouTube iframes with a flat link to the page.
+    #
+    # Params:
+    # +html+:: HTML string to clean.
+    #
+    def fix_youtube_iframes(html)
+      doc = Nokogiri::HTML(html)
+      doc.search('//iframe').each do |f_node|
+        video_id = f_node.attributes['id'].to_s.split('_').last
+        url = "https://youtube.com/watch?v=#{video_id}"
+        new_node = Nokogiri::HTML.fragment("<p><a href=\"#{url}\">#{url}</a></p>")
+        f_node.replace(new_node)
+      end
+      doc.at_xpath('//body').to_s[('<body>'.length)..-('</body>'.length + 1)]
+    end
+
     def pretty_print_xml(xml)
       doc = Nokogiri.XML(xml) do |config|
         config.default_xml.noblanks
