+++ src/_posts/2018/2018-09-22-suspicious-minds.md
@@ -26,7 +26,11 @@ This is my abstract:
 
 The talk was recorded, and thanks to the wizardry of Tim&nbsp;Williams, it was on YouTube within a day:
 
-{% youtube https://www.youtube.com/watch?v=B3XxPnbehqQ %}
+{%
+  include embeds/youtube.html
+  url="https://www.youtube.com/watch?v=B3XxPnbehqQ"
+  title="Building trust in an age of suspicious minds"
+%}
 
 You can read the slides and transcript on this page, or download the slides [as a PDF](/files/2018/suspicious-minds.pdf).
 The transcript is based on the captions on the YouTube video, with some light editing and editorial notes where required.
