+++ src/_posts/2019/2019-01-31-monki-gras-the-curb-cut-effect.md
@@ -23,7 +23,11 @@ It was a lot of fun to write and present.
 
 The talk was recorded, and you can watch it on YouTube:
 
-{% youtube https://www.youtube.com/watch?v=6Yf3iukx3tQ %}
+{%
+  include embeds/youtube.html
+  url="https://www.youtube.com/watch?v=6Yf3iukx3tQ"
+  title="The Curb Cut Effect"
+%}
 
 You can read the slides and my notes on this page, or download the slides [as a PDF](/files/2019/curbcut-monkigras.pdf).
